Fix group permission handling
The logic to determine whether a connected WebSocket client had permission to
read a particular annotation was broken because the `AnnotationContext.__acl__`
method has an implicit dependency on the current Pyramid registry and this
registry was never set in the WS server. As a result, the call to
`principals_allowed_by_permission` in `AnnotationContext.__acl__` ended up using
the default authorization policy which always returns `[system.Everyone]` - ie.
as if any connected WS client could read the annotation.

This commit fixes the issue by using a helper from `pyramid.scripting` to set
the active registry while generating the notification. It also simplifies the
logic to determine the principals (groups, users etc.) that can read the
annotation and compare them to the principals of the WebSocket client.